Fix a RedScreen during Sharpshooter's Faceoff ability
The root cause is the X2Effect_ApplyWeaponDamage instances that
we replace with an _RD version are still actually present in the list
of TargetEffects of an ability. We cannot remove them from this list as
it is protectedwrite. We have disabled these effects but apparently
they still do some work during visualization.