No longer exclude ApplyWeaponDamage effects with bIgnoreBaseDamage=true
This created a compatibility issue with the mod [WOTC] Hand to Hand
Abilities which for some reason set that flag to true on their custom
abilities. We excluded abilities like that because this flag seemingly
was set on abilities that we don't really need to touch such as
HomingMineDetonation. However, we only excluded 4 abilities total
because of it and enabling this mod for them does not seem to hurt.
If it turns out to be an issue we can use the new config
DisableForAbilities to exclude them by default in the mod.