Fix Lightning Reflexes not triggering when initial roll was a miss
When the game rolls a miss it does not apply Lightning Reflexes which
makes sense for the base game. However, this means our mod will then
turn that miss (which is reported as a regular miss) into a hit with
altered damage. That is not supposed to happen, the first shot should
simply miss in our mod as well (and be tagged as a "Lightning Reflexes"
miss). This is now fixed.